+++ VpnHood.Client.App.UI.Html/src/pages/Home.vue
@@ -64,6 +64,7 @@
           <ClientProfileMenu
             clientProfileId="$"
             color="white"
+            :showSettingsItem="true"
             :showAddServerItem="true"
             :showDeleteItem="false"
             :showRenameItem="false"